Add llvm13 extension
Some C++ projects use clang-format in a pre-commit hook. The freedesktop
SDK used to contain clang and clang-format, but now it lives in an
extension. This change mounts the extension in the app, and sets up the
environment variables so that scripts can override the path for
clang-format and clang-format-diff.

Closes: #32